Separate analyzers and extractors
